refactor: Remove pkg directories, move packages to appropriate locations
Move agent and CNI packages out of pkg/ directories:
- agent/pkg/cmd → agent/internal/cmd
- agent/pkg/constants → agent/constants
- agent/pkg/storage → agent/storage
- agent/pkg/types → agent/types
- cni/pkg/cmd → cni/internal/cmd
- cni/pkg/config → cni/config
