+++ agent/internal/cmd/BUILD.bazel
@@ -6,16 +6,16 @@ go_library(
         "config.go",
         "root.go",
     ],
-    importpath = "github.com/bpalermo/aether/agent/pkg/cmd",
+    importpath = "github.com/bpalermo/aether/agent/internal/cmd",
     visibility = ["//visibility:public"],
     deps = [
+        "//agent/constants",
         "//agent/internal/awsconfig",
         "//agent/internal/cni/server",
         "//agent/internal/spire",
         "//agent/internal/xds/cache",
         "//agent/internal/xds/server",
-        "//agent/pkg/constants",
-        "//agent/pkg/storage",
+        "//agent/storage",
         "//api/aether/cni/v1:cni",
         "//common/must",
         "//log",
@@ -31,8 +31,8 @@ go_test(
     srcs = ["config_test.go"],
     embed = [":cmd"],
     deps = [
+        "//agent/constants",
         "//agent/internal/cni/server",
-        "//agent/pkg/constants",
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
     ],
