+++ agent/internal/cni/server/BUILD.bazel
@@ -10,13 +10,13 @@ go_library(
     importpath = "github.com/bpalermo/aether/agent/internal/cni/server",
     visibility = ["//agent:__subpackages__"],
     deps = [
+        "//agent/constants",
         "//agent/internal/envoy/admin",
         "//agent/internal/spire",
         "//agent/internal/xds/cache",
         "//agent/internal/xds/proxy",
-        "//agent/pkg/constants",
-        "//agent/pkg/storage",
-        "//agent/pkg/types",
+        "//agent/storage",
+        "//agent/types",
         "//api/aether/cni/v1:cni",
         "//constants",
         "//registry",
@@ -41,12 +41,12 @@ go_test(
     ],
     embed = [":server"],
     deps = [
+        "//agent/constants",
         "//agent/internal/envoy/admin",
         "//agent/internal/spire",
         "//agent/internal/xds/cache",
-        "//agent/pkg/constants",
-        "//agent/pkg/storage",
-        "//agent/pkg/types",
+        "//agent/storage",
+        "//agent/types",
         "//api/aether/cni/v1:cni",
         "//api/aether/registry/v1:registry",
         "//constants",
@@ -81,9 +81,9 @@ go_test(
         "//agent/internal/envoy/admin",
         "//agent/internal/spire",
         "//agent/internal/xds/cache",
-        "//agent/pkg/constants",
-        "//agent/pkg/storage",
-        "//agent/pkg/types",
+        "//agent/constants",
+        "//agent/storage",
+        "//agent/types",
         "//api/aether/cni/v1:cni",
         "//api/aether/registry/v1:registry",
         "//constants",
