+++ agent/internal/cni/server/pod_test.go
@@ -11,9 +11,9 @@ import (
 	"github.com/bpalermo/aether/agent/internal/envoy/admin"
 	"github.com/bpalermo/aether/agent/internal/spire"
 	"github.com/bpalermo/aether/agent/internal/xds/cache"
-	agentconstants "github.com/bpalermo/aether/agent/pkg/constants"
-	"github.com/bpalermo/aether/agent/pkg/storage"
-	"github.com/bpalermo/aether/agent/pkg/types"
+	agentconstants "github.com/bpalermo/aether/agent/constants"
+	"github.com/bpalermo/aether/agent/storage"
+	"github.com/bpalermo/aether/agent/types"
 	cniv1 "github.com/bpalermo/aether/api/aether/cni/v1"
 	registryv1 "github.com/bpalermo/aether/api/aether/registry/v1"
 	"github.com/bpalermo/aether/constants"
