+++ agent/internal/cni/server/server_integration_test.go
@@ -9,11 +9,11 @@ import (
 	"testing"
 	"time"
 
-	agentconstants "github.com/bpalermo/aether/agent/pkg/constants"
+	agentconstants "github.com/bpalermo/aether/agent/constants"
 	"github.com/bpalermo/aether/agent/internal/spire"
 	"github.com/bpalermo/aether/agent/internal/xds/cache"
-	"github.com/bpalermo/aether/agent/pkg/storage"
-	"github.com/bpalermo/aether/agent/pkg/types"
+	"github.com/bpalermo/aether/agent/storage"
+	"github.com/bpalermo/aether/agent/types"
 	cniv1 "github.com/bpalermo/aether/api/aether/cni/v1"
 	"github.com/bpalermo/aether/constants"
 	"github.com/go-logr/logr"
