+++ agent/internal/xds/server/BUILD.bazel
@@ -6,9 +6,9 @@ go_library(
     importpath = "github.com/bpalermo/aether/agent/internal/xds/server",
     visibility = ["//agent:__subpackages__"],
     deps = [
+        "//agent/constants",
         "//agent/internal/xds/cache",
-        "//agent/pkg/constants",
-        "//agent/pkg/storage",
+        "//agent/storage",
         "//api/aether/cni/v1:cni",
         "//registry",
         "//xds",
@@ -26,7 +26,7 @@ go_test(
     deps = [
         "//agent/internal/xds/cache",
         "//agent/internal/xds/proxy",
-        "//agent/pkg/storage",
+        "//agent/storage",
         "//api/aether/cni/v1:cni",
         "//api/aether/registry/v1:registry",
         "//registry",
@@ -58,7 +58,7 @@ go_test(
         "//agent/internal/xds/cache",
         "//agent/internal/xds/config",
         "//agent/internal/xds/proxy",
-        "//agent/pkg/storage",
+        "//agent/storage",
         "//api/aether/cni/v1:cni",
         "//api/aether/registry/v1:registry",
         "//registry",
