+++ agent/storage/mock.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"errors"
 
-	"github.com/bpalermo/aether/agent/pkg/types"
+	"github.com/bpalermo/aether/agent/types"
 	"google.golang.org/protobuf/proto"
 )
