+++ agent/types/BUILD.bazel
@@ -3,6 +3,6 @@ load("@rules_go//go:def.bzl", "go_library")
 go_library(
     name = "types",
     srcs = ["container.go"],
-    importpath = "github.com/bpalermo/aether/agent/pkg/types",
+    importpath = "github.com/bpalermo/aether/agent/types",
     visibility = ["//visibility:public"],
 )
