+++ cni/internal/install/cniconfig.go
@@ -9,9 +9,9 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/bpalermo/aether/agent/pkg/constants"
+	"github.com/bpalermo/aether/agent/constants"
 	"github.com/bpalermo/aether/cni/internal/util"
-	"github.com/bpalermo/aether/cni/pkg/config"
+	"github.com/bpalermo/aether/cni/config"
 	"github.com/bpalermo/aether/common/file"
 	"github.com/containernetworking/cni/libcni"
 )
